Barrier has been discovered!
Barriers shrink in size when a alien laser hits it. 10 hits and the barrier is destroyed. the 10 hits can be altered in settings. a commented function called regrow is also there to allow ship lasers to regrow the barrier back to full health. I add this is the window version makes every thing to fast to play normal space invaders so I added it as a way to balence it out better.